refactor(stripe): consolidate Stripe service methods into a single stripeService for improved organization and clarity
