+++ web-app/src/lib/clients/stripe.client.ts
@@ -1,3 +1,5 @@
+import "server-only";
+
 import { headers } from "next/headers";
 import Stripe from "stripe";
