+++ web-app/src/lib/services/index.ts
@@ -4,6 +4,6 @@ export * from "./invitation";
 export * from "./job";
 export * from "./lock.service";
 export * from "./organization";
-export * from "./stripe";
+export * from "./stripe.service";
 export * from "./user.service";
 export * from "./utm.service";
