--- web-app/src/lib/services/stripe/index.ts
@@ -1,2 +1,0 @@
-export * from "./service";
-export * from "./third-party";
